center stage & chore
